Added source files
The added source files are incomplete and have lot of issues.
Their goals was not met and it seems impossible to accomplish it with their implementation.
The goal of the project is not realistic and its stuck between multiple libraries.

Most problems come with cross library use including 'requests' and 'aiohttp'.
Creating generic bruteforce library for website does not seem realistic.
There are lot of things that can be done and its hard to represent them all.

This is a failure in terms of idea and implementation.
Unless the idea is modified, there wont be any fruits from it.
Its better to choose one library and continue working on it.

Another way could be creating wrapper library that is built on top of the libraries.
Most lines of code was trying to ensure code works in both 'requests' and 'aiohttp'.

Its better to use 'broote' direcly which is more straight forward and good api.