Updated 'broote' version(#3)
Broote v0.4.0 introdused incompatible changes which also affected realsed version of 'webrute'.
It was unnexpected that 'broote' would release breaking release in short time.
'webrute' expected it to happen at version 1.0.0 but was wrong.

This commit contain code from webrute v0.0.1 with version for 'webrute' changed to exclude v0.4.0 and more.
Code in this commit is not available in latest commit but it also impacted by this incompatibily issue.