Direct session no longer closed
Session that is provided directly is no longer closed.
When session is provided direcly it means it will be reused in multiple places.
This would allow multiple runners to share the same session without one of them closing it.

Session is means to be shared not just be closed by single runner.
But providing session as callable would mean it will be created and closed internally.